requirements.txt
streamlit
google.generativeai
dotenv